Replace the player age with year of birth

+++ laravel/database/migrations/2026_05_04_184249_create_players_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->id();
             $table->string('firstname');
             $table->string('lastname');
-            $table->unsignedTinyInteger('year_of_birth');
+            $table->unsignedSmallInteger('year_of_birth');
             $table->foreignIdFor(Club::class)->constrained()->cascadeOnUpdate()->restrictOnDelete();
         });
